Only dispatch on first argument
I think this is correct because Pool doesn't involve itself in any of the arugments; it just checks out the object then re-dispatches. This is potentially a breaking change because it could introduce new method dispatch ambiguities, but I don't think that's a concern here because Pool has neither superclasses or subclasses.

Fixes #120